Improved multi-server support for columnar storage

+++ src/main/java/net/coreprotect/database/Database.java
@@ -30,6 +30,7 @@ import net.coreprotect.consumer.process.Process;
 import net.coreprotect.database.clickhouse.ClickHouseConsumerWriteBatch;
 import net.coreprotect.database.clickhouse.ClickHouseDatabase;
 import net.coreprotect.database.clickhouse.ClickHouseJdbcConfig;
+import net.coreprotect.database.statement.UserStatement;
 import net.coreprotect.language.Phrase;
 import net.coreprotect.listener.player.InventoryChangeListener;
 import net.coreprotect.model.BlockGroup;
@@ -598,6 +599,18 @@ public class Database extends Queue {
         return new RelationalConsumerWriteBatch(connection, ConfigHandler.databaseType);
     }
 
+    public static int nextClickHouseIdentifierId(ConsumerWriteBatch.ReferenceKind kind, String value, int currentMaximum) throws SQLException {
+        return requireClickHouseDatabase().nextIdentifierId(kind, value, currentMaximum);
+    }
+
+    public static String findClickHouseIdentifierValue(ConsumerWriteBatch.ReferenceKind kind, int id) throws SQLException {
+        return requireClickHouseDatabase().findIdentifierValue(kind, id);
+    }
+
+    public static int findClickHouseIdentifierId(ConsumerWriteBatch.ReferenceKind kind, String value) throws SQLException {
+        return requireClickHouseDatabase().findIdentifierId(kind, value);
+    }
+
     public static void recordDatabaseVersion(Statement statement, String version) throws SQLException {
         if (ConfigHandler.databaseType.isClickHouse()) {
             requireClickHouseDatabase().updateCoreVersion(version);
@@ -608,6 +621,9 @@ public class Database extends Queue {
     }
 
     public static long purgeClickHouse(long startTime, long endTime, int worldId, List<Integer> blockTypes, boolean optimize) throws SQLException {
+        if (!Config.getGlobal().DATABASE_LOCK) {
+            throw new SQLException("ClickHouse purge requires database-lock to be enabled and every other CoreProtect installation sharing the database and prefix to be stopped");
+        }
         return requireClickHouseDatabase().purge(startTime, endTime, worldId, blockTypes, optimize);
     }
 
@@ -1186,6 +1202,7 @@ public class Database extends Queue {
     private static synchronized void closeClickHouseChecked() throws SQLException {
         ClickHouseDatabase database = clickHouseDatabase;
         clickHouseDatabase = null;
+        UserStatement.clearClickHouseCaches();
         if (database != null) {
             database.close();
         }
